requirements.txt
pytest==3.8.0
PyYAML==3.13
